:paperclip: defensive parser invocation

+++ lib/metastatic/adapters/haskell/subprocess.ex
@@ -20,27 +20,31 @@ defmodule Metastatic.Adapters.Haskell.Subprocess do
   """
   @spec parse(String.t()) :: {:ok, map()} | {:error, String.t()}
   def parse(source) when is_binary(source) do
-    temp_path =
-      Path.join(
-        System.tmp_dir!(),
-        "metastatic_hs_#{System.unique_integer([:positive, :monotonic])}.txt"
-      )
+    if File.dir?(@parser_dir) and System.find_executable("stack") != nil do
+      temp_path =
+        Path.join(
+          System.tmp_dir!(),
+          "metastatic_hs_#{System.unique_integer([:positive, :monotonic])}.txt"
+        )
 
-    File.write!(temp_path, source)
+      File.write!(temp_path, source)
 
-    try do
-      case System.cmd("stack", ["exec", "parser", "--", temp_path],
-             cd: @parser_dir,
-             stderr_to_stdout: true
-           ) do
-        {output, 0} ->
-          handle_parser_output(output)
+      try do
+        case System.cmd("stack", ["exec", "parser", "--", temp_path],
+               cd: @parser_dir,
+               stderr_to_stdout: true
+             ) do
+          {output, 0} ->
+            handle_parser_output(output)
 
-        {error_output, _exit_code} ->
-          {:error, "Haskell parser failed: #{inspect(error_output)}"}
+          {error_output, _exit_code} ->
+            {:error, "Haskell parser failed: #{inspect(error_output)}"}
+        end
+      after
+        File.rm(temp_path)
       end
-    after
-      File.rm(temp_path)
+    else
+      {:error, "Haskell parser unavailable"}
     end
   end
